# backend/debug_openapi.json
{"openapi":"3.1.0","info":{"title":"AUTOPILOT API","description":"IA d'automatisation de processus métier","version":"1.0.0"},"paths":{"/api/v1/contract/analyze":{"post":{"tags":["Contract Reader"],"summary":"Analyze Contract","description":"Endpoint simplifié pour analyse de contrat - Version test sans GDPR","operationId":"analyze_contract_api_v1_contract_analyze_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyze_contract_api_v1_contract_analyze_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/contract/download/{file_id}":{"get":{"tags":["Contract Reader"],"summary":"Download Summary Pdf","description":"Télécharge le PDF d'un résumé avec URL signée","operationId":"download_summary_pdf_api_v1_contract_download__file_id__get","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}},{"name":"t","in":"query","required":true,"schema":{"type":"integer","title":"T"}},{"name":"s","in":"query","required":true,"schema":{"type":"string","title":"S"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/contract/download-pdf":{"post":{"tags":["Contract Reader"],"summary":"Download Contract Pdf","description":"Génère et télécharge le PDF du résumé de contrat\nExclut les métadonnées techniques pour un document professionnel","operationId":"download_contract_pdf_api_v1_contract_download_pdf_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_download_contract_pdf_api_v1_contract_download_pdf_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/contract/consent":{"post":{"tags":["Contract Reader"],"summary":"Record User Consent","description":"Enregistre le consentement GDPR de l'utilisateur","operationId":"record_user_consent_api_v1_contract_consent_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"boolean"},"type":"object","title":"Consents"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/contract/gdpr/erase":{"delete":{"tags":["Contract Reader"],"summary":"Request Data Erasure","description":"Demande d'effacement GDPR (droit à l'oubli)","operationId":"request_data_erasure_api_v1_contract_gdpr_erase_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/contract/health":{"get":{"tags":["Contract Reader"],"summary":"Health Check","description":"Vérification de l'état du service","operationId":"health_check_api_v1_contract_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/contract/stats":{"get":{"tags":["Contract Reader"],"summary":"Get Stats","description":"Statistiques détaillées du système","operationId":"get_stats_api_v1_contract_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/contract/budget":{"get":{"tags":["Contract Reader"],"summary":"Get Budget Status","description":"Statut du budget utilisateur","operationId":"get_budget_status_api_v1_contract_budget_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/contract/test":{"post":{"tags":["Contract Reader Simple"],"summary":"Test Contract Analysis","description":"Endpoint de test simple pour analyse de contrat - Pas de GDPR","operationId":"test_contract_analysis_api_v1_contract_test_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_test_contract_analysis_api_v1_contract_test_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","description":"Root endpoint","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Root  Get"}}}}}}},"/api/v1/health":{"get":{"summary":"Health Check","description":"Health check basique","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/api/v1/ready":{"get":{"summary":"Readiness Check","description":"Readiness check avec dépendances","operationId":"readiness_check_api_v1_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/metrics":{"get":{"summary":"Prometheus Metrics","description":"Métriques Prometheus","operationId":"prometheus_metrics_api_v1_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/extract":{"post":{"summary":"Extract Document","description":"Extraction de document avec IA\n- Validation sécurité\n- Cache intelligent\n- Création ticket Jira\n- Journalisation complète","operationId":"extract_document_api_v1_extract_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_extract_document_api_v1_extract_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExtractionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/quality/dashboard":{"get":{"summary":"Quality Dashboard","description":"Dashboard qualité temps réel","operationId":"quality_dashboard_api_v1_quality_dashboard_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/demo/reset":{"post":{"summary":"Reset Demo","description":"Reset environnement démo (idempotent)","operationId":"reset_demo_api_v1_demo_reset_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_analyze_contract_api_v1_contract_analyze_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"summary_mode":{"type":"string","title":"Summary Mode","default":"standard"}},"type":"object","required":["file"],"title":"Body_analyze_contract_api_v1_contract_analyze_post"},"Body_download_contract_pdf_api_v1_contract_download_pdf_post":{"properties":{"processing_id":{"type":"string","title":"Processing Id"},"user_id":{"type":"string","title":"User Id"}},"type":"object","required":["processing_id","user_id"],"title":"Body_download_contract_pdf_api_v1_contract_download_pdf_post"},"Body_extract_document_api_v1_extract_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_extract_document_api_v1_extract_post"},"Body_test_contract_analysis_api_v1_contract_test_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"summary_mode":{"type":"string","title":"Summary Mode","default":"standard"}},"type":"object","required":["file"],"title":"Body_test_contract_analysis_api_v1_contract_test_post"},"ContractExtraction":{"properties":{"contract_type":{"$ref":"#/components/schemas/ContractType"},"parties":{"items":{"type":"string"},"type":"array","maxItems":10,"minItems":1,"title":"Parties"},"amount":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Amount"},"currency":{"anyOf":[{"type":"string","maxLength":3},{"type":"null"}],"title":"Currency"},"start_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Start Date"},"end_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"End Date"},"key_terms":{"items":{"type":"string"},"type":"array","maxItems":20,"title":"Key Terms"},"summary":{"type":"string","maxLength":1000,"minLength":10,"title":"Summary"},"confidence_score":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Confidence Score"},"extracted_fields":{"additionalProperties":true,"type":"object","title":"Extracted Fields"}},"type":"object","required":["contract_type","parties","summary","confidence_score"],"title":"ContractExtraction","description":"Résultat d'extraction de contrat"},"ContractType":{"type":"string","enum":["service","purchase","employment","lease","other"],"title":"ContractType","description":"Types de contrats supportés"},"ExtractionResponse":{"properties":{"session_id":{"type":"string","title":"Session Id"},"extraction":{"$ref":"#/components/schemas/ContractExtraction"},"jira_ticket":{"anyOf":[{"$ref":"#/components/schemas/JiraTicket"},{"type":"null"}]},"processing_time_ms":{"type":"integer","title":"Processing Time Ms"},"cached":{"type":"boolean","title":"Cached","default":false},"demo_mode":{"type":"boolean","title":"Demo Mode","default":true},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"}},"type":"object","required":["session_id","extraction","processing_time_ms"],"title":"ExtractionResponse","description":"Réponse complète d'extraction"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status","default":"healthy"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"demo_mode":{"type":"boolean","title":"Demo Mode","default":true},"version":{"type":"string","title":"Version","default":"1.0.0"}},"type":"object","required":["timestamp"],"title":"HealthResponse","description":"Réponse health check"},"JiraTicket":{"properties":{"key":{"type":"string","title":"Key"},"url":{"type":"string","title":"Url"},"summary":{"type":"string","title":"Summary"},"description":{"anyOf":[{"type":"string"},{"additionalProperties":true,"type":"object"}],"title":"Description"},"project_key":{"type":"string","title":"Project Key"},"issue_type":{"type":"string","title":"Issue Type","default":"Task"},"priority":{"type":"string","title":"Priority","default":"Medium"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"demo_mode":{"type":"boolean","title":"Demo Mode","default":false}},"type":"object","required":["key","url","summary","description","project_key","created_at"],"title":"JiraTicket","description":"Ticket Jira créé"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}